modified train controllers and reservation controllers

+++ TicketingSystem.Repository/ReservationService.cs
@@ -24,12 +24,14 @@ namespace TicketingSystem.Repository
 
         }
 
-        public BaseResponse InsertReservation(Reservation Reservation)
+        public BaseResponse InsertReservation(Reservation reservation)
         {
             try
-            {                
-                _reservationCollection.InsertOne(Reservation);
-                return new BaseResponseService().GetSuccessResponse(Reservation);
+            {
+                Reservation resByUserId = new Reservation();
+                //resByUserId = _reservationCollection.Find(u => u.UserId == reservation.UserId).SingleOrDefault();
+                _reservationCollection.InsertOne(reservation);
+                return new BaseResponseService().GetSuccessResponse(reservation);
             }
             catch (Exception ex)
             {
@@ -42,26 +44,28 @@ namespace TicketingSystem.Repository
         {
             try
             {
-                var reservations = _reservationCollection.Find(_ => true).ToList();
-                var reservationList = new List<object>();
-
-                foreach (var reservation in reservations)
+                var reservations = _reservationCollection.Find( _ => true).ToList();
+                if (reservations != null)
                 {
-                    User user = _userCollection.Find(u => u.Id == reservation.UserId).SingleOrDefault();
-                    Train train = _trainCollection.Find(t => t.Id == reservation.TrainId).SingleOrDefault();
+                    var reservationList = new List<object>();
 
-                    var resObject = new { reservation, user, train };
-                    reservationList.Add(resObject);
+                    foreach (var reservation in reservations)
+                    {
+                        User user = _userCollection.Find(u => u.Id == reservation.UserId).SingleOrDefault();
+                        Train train = _trainCollection.Find(t => t.Id == reservation.TrainId).SingleOrDefault();
+
+                        var resObject = new { reservation, user, train };
+                        reservationList.Add(resObject);
+                    }
+                    return new BaseResponseService().GetSuccessResponse(reservationList);
                 }
-
-                return new BaseResponseService().GetSuccessResponse(reservationList);
+                return new BaseResponseService().GetValidatationResponse("Reservations Not Found!");
             }
             catch (Exception ex)
             {
                 return new BaseResponseService().GetErrorResponse(ex);
             }
         }
-
 
 
         public BaseResponse GetReservationById(string id)
@@ -72,13 +76,14 @@ namespace TicketingSystem.Repository
                 User user = new User();
                 Train train = new Train();
                 reservation = _reservationCollection.Find(s => s.Id == id).SingleOrDefault();
-                user =  _userCollection.Find(t => t.Id == reservation.UserId).SingleOrDefault();
-                train = _trainCollection.Find(t => t.Id == reservation.TrainId).SingleOrDefault();
-
-                var resObject = new { reservation = reservation, user = user, train = train };
-
-                return new BaseResponseService().GetSuccessResponse(resObject);
-
+                if(reservation != null)
+                {
+                    user = _userCollection.Find(t => t.Id == reservation.UserId).SingleOrDefault();
+                    train = _trainCollection.Find(t => t.Id == reservation.TrainId).SingleOrDefault();
+                    var resObject = new { reservation = reservation, user = user, train = train };
+                    return new BaseResponseService().GetSuccessResponse(resObject);
+                }
+                return new BaseResponseService().GetValidatationResponse("Reservation Not Found!");
             }
             catch (Exception ex)
             {
@@ -92,18 +97,27 @@ namespace TicketingSystem.Repository
             try
             {
                 var reservations = _reservationCollection.Find(r => r.UserId == userId).ToList();
-                var reservationList = new List<object>();
-
-                foreach (var reservation in reservations)
+                if(reservations != null)
                 {
-                    User user = _userCollection.Find(u => u.Id == reservation.UserId).SingleOrDefault();
-                    Train train = _trainCollection.Find(t => t.Id == reservation.TrainId).SingleOrDefault();
+                    if(reservations[0].UserId != null)
+                    {
+                        var reservationList = new List<object>();
 
-                    var resObject = new { reservation, user, train };
-                    reservationList.Add(resObject);
+                        foreach (var reservation in reservations)
+                        {
+                            User user = _userCollection.Find(u => u.Id == reservation.UserId).SingleOrDefault();
+                            Train train = _trainCollection.Find(t => t.Id == reservation.TrainId).SingleOrDefault();
+
+                            var resObject = new { reservation, user, train };
+                            reservationList.Add(resObject);
+                        }
+
+                        return new BaseResponseService().GetSuccessResponse(reservationList);
+                    }
+                    return new BaseResponseService().GetValidatationResponse("User Not Found!");
                 }
+                return new BaseResponseService().GetValidatationResponse("User dose not have any Reservation!");
 
-                return new BaseResponseService().GetSuccessResponse(reservationList);
             }
             catch (Exception ex)
             {
@@ -116,12 +130,18 @@ namespace TicketingSystem.Repository
         {
             try
             {
-                var filter = Builders<Reservation>.Filter.Eq(s => s.Id, id);
-                var update = Builders<Reservation>.Update
-                    .Set(s => s.Status, updatedReservation.Status);
+                Reservation reservation = new Reservation();
+                reservation = _reservationCollection.Find(s => s.Id == id).SingleOrDefault();
+                if (reservation != null)
+                {
+                    var filter = Builders<Reservation>.Filter.Eq(s => s.Id, id);
+                    var update = Builders<Reservation>.Update.Set(s => s.Status, updatedReservation.Status);
 
-                _reservationCollection.UpdateOne(filter, update);
-                return new BaseResponseService().GetSuccessResponse(updatedReservation);
+                    _reservationCollection.UpdateOne(filter, update);
+                    return new BaseResponseService().GetSuccessResponse(updatedReservation);
+                }
+                return new BaseResponseService().GetValidatationResponse("Reservation Not Found!");
+
             }
             catch (Exception ex)
             {
@@ -134,10 +154,15 @@ namespace TicketingSystem.Repository
         {
             try
             {
-                var filter = Builders<Reservation>.Filter.Eq(s => s.Id, id);
-                _reservationCollection.DeleteOne(filter);
-                return new BaseResponseService().GetSuccessResponse(filter);
-
+                Reservation reservation = new Reservation();
+                reservation = _reservationCollection.Find(s => s.Id == id).SingleOrDefault();
+                if (reservation != null)
+                {
+                    var filter = Builders<Reservation>.Filter.Eq(s => s.Id, id);
+                    _reservationCollection.DeleteOne(filter);
+                    return new BaseResponseService().GetSuccessResponse("Reservation Deleted Successfully");
+                }
+                return new BaseResponseService().GetValidatationResponse("Reservation Not Found!");
             }
             catch (Exception ex)
             {
